worker/requirements.txt
pika
pandas
scikit-learn
joblib
nltk
pydantic
psycopg2-binary